minimum working example of image plot
