+++ hp-mri-frontend/src/components/visualize/PlotShiftPanel.tsx
@@ -41,86 +41,84 @@ const PlotShiftPanel: React.FC<Props> = ({
                 zIndex: 1302,
             }}
         >
-            {mode === 'imaging' && (
-                <Box sx={{ width: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
-                    <Box
-                        sx={{
-                            position: 'absolute',
-                            top: -150,
-                            left: 0,
-                            width: 150,
-                            height: 140,
-                            overflowY: 'auto',
-                            border: '1px solid #333',  // darker subtle border
-                            borderRadius: 2,
-                            backgroundColor: '#1e1e1e',  // match PlotShiftPanel
-                            color: 'white',              // white text
-                            padding: 1,
-                            zIndex: 5,
-                            boxShadow: 4,                // same boxShadow
-                        }}
-                        tabIndex={0}
-                    >
-                        <Typography align="center" sx={{ mb: 1, fontSize: 14, color: 'white' }}>
-                            Metabolite
-                        </Typography>
-                        {metaboliteNames.map((name, i) => (
-                            <Box
-                                key={i}
-                                onClick={() => { onMetaboliteChange(i); setMetaboliteIndex(i); }}
-                                onDoubleClick={() => setEditingIndex(i)}
-                                sx={{
-                                    paddingY: 0.5,
-                                    paddingX: 1,
-                                    cursor: 'pointer',
-                                    fontSize: 13,
-                                    backgroundColor: metaboliteIndex === i ? '#1976d2' : 'transparent',
-                                    color: metaboliteIndex === i ? '#fff' : '#ddd',
-                                    borderRadius: 1,
-                                    '&:hover': {
-                                        backgroundColor: metaboliteIndex === i ? '#1565c0' : '#2e2e2e',
-                                    },
-                                }}
-                            >
-                                {editingIndex === i ? (
-                                    <TextField
-                                        autoFocus
-                                        variant="standard"
-                                        value={name}
-                                        onChange={(e) => {
-                                            const updated = [...metaboliteNames];
-                                            updated[i] = e.target.value;
-                                            setMetaboliteNames(updated);
-                                        }}
-                                        onBlur={() => setEditingIndex(null)}
-                                        onKeyDown={(e) => {
-                                            if (e.key === 'Enter') setEditingIndex(null);
-                                        }}
-                                        InputProps={{
-                                            disableUnderline: true,
-                                            style: {
-                                                fontSize: 13,
-                                                color: 'white',
-                                                padding: 0,
-                                            },
-                                        }}
-                                        sx={{
-                                            input: {
-                                                padding: 0,
-                                                color: 'black',
-                                                backgroundColor: 'transparent',
-                                                textAlign: 'center',
-                                            },
-                                        }}
-                                    />
-                                ) : (
-                                    name
-                                )}
-                            </Box>
-                        ))}
-                    </Box>
+            <Box sx={{ width: '100%', display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+                <Box
+                    sx={{
+                        position: 'absolute',
+                        top: -150,
+                        left: 0,
+                        width: 150,
+                        height: 140,
+                        overflowY: 'auto',
+                        border: '1px solid #333',  // darker subtle border
+                        borderRadius: 2,
+                        backgroundColor: '#1e1e1e',  // match PlotShiftPanel
+                        color: 'white',              // white text
+                        padding: 1,
+                        zIndex: 5,
+                        boxShadow: 4,                // same boxShadow
+                    }}
+                    tabIndex={0}
+                >
+                    <Typography align="center" sx={{ mb: 1, fontSize: 14, color: 'white' }}>
+                        Metabolite
+                    </Typography>
+                    {metaboliteNames.map((name, i) => (
+                        <Box
+                            key={i}
+                            onClick={() => { onMetaboliteChange(i); setMetaboliteIndex(i); }}
+                            onDoubleClick={() => setEditingIndex(i)}
+                            sx={{
+                                paddingY: 0.5,
+                                paddingX: 1,
+                                cursor: 'pointer',
+                                fontSize: 13,
+                                backgroundColor: metaboliteIndex === i ? '#1976d2' : 'transparent',
+                                color: metaboliteIndex === i ? '#fff' : '#ddd',
+                                borderRadius: 1,
+                                '&:hover': {
+                                    backgroundColor: metaboliteIndex === i ? '#1565c0' : '#2e2e2e',
+                                },
+                            }}
+                        >
+                            {editingIndex === i ? (
+                                <TextField
+                                    autoFocus
+                                    variant="standard"
+                                    value={name}
+                                    onChange={(e) => {
+                                        const updated = [...metaboliteNames];
+                                        updated[i] = e.target.value;
+                                        setMetaboliteNames(updated);
+                                    }}
+                                    onBlur={() => setEditingIndex(null)}
+                                    onKeyDown={(e) => {
+                                        if (e.key === 'Enter') setEditingIndex(null);
+                                    }}
+                                    InputProps={{
+                                        disableUnderline: true,
+                                        style: {
+                                            fontSize: 13,
+                                            color: 'white',
+                                            padding: 0,
+                                        },
+                                    }}
+                                    sx={{
+                                        input: {
+                                            padding: 0,
+                                            color: 'black',
+                                            backgroundColor: 'transparent',
+                                            textAlign: 'center',
+                                        },
+                                    }}
+                                />
+                            ) : (
+                                name
+                            )}
+                        </Box>
+                    ))}
                 </Box>
-            )}
+            </Box>
 
             <Box display="flex" justifyContent="center">
                 <Tooltip title="Up">
